migrating from prisma to kysely
prisma has proven too difficult to maintain. prisma is a good ORM
providing typesafety database calls. however, prisma's approach to
building a client has made it extremely difficult to embed in multiple
projects. the focus of prisma seems to be on server side databases
embedded into a single cloud application

kysely is a query builder that also provides type safety. however, due
to its simplicity and lack of a separate client generation process, it's
much easier to embed into both a webapp and a CLI.

this migration maintains typesafety when it comes to database calls and
eases the build steps for both web/cli packages.