adding in the second piece: link
a lot of reworking for what is a piece and what should be abstracted out
to individual pieces vs pushing the logic down into the piece itself

we still are missing fetch and attach functionality for links, but this
is getting large as it is

one big jump is a custom AJV jtd schema constraint to allow for a
frontmatter field to only contain a date string. AJV advises against
custom formats for JTD as they don't work across languages and are
outside of the spec.

i think it is ok, but am now interested in moving more schema stuff into
the kysely package itself. (though it will no longer just be a DB
package, but more of a schema/format representation package).