migrate core from JTD to JSONSchema
in an effort to move away from hardcoded metadata that only effects
typings, we migrate to json schema to be better aligned with the intent
on modifying data and formats, not just types.

json schema also has a path to appending new formats to add more complex
validations to the data layer itself

used this opportunity to simplify reading and parsing of the schema and
pulling up all the exports into the main library

bump up typescript, kysely and eslint as well

lastly, continue pushing all piece abstractions as low as possible to
minimize the amount of work needed to introduce new pieces. with the
eventual goal of new piece introduction purely with a jsonschema (maybe
in yaml)