[cli] clear piece cache when updating piece schema
when a schema is updated, we recreate the piece table. but since it is
now empty, we also need to re-add all pieces back to re-populate the
table

however, the cache is still comparing the files to their previous
version and thus syncing won't re-add them. if we clear the cache of all
the piece types, a sync will now work properly