[cli] properly parse csv and file attachments
(*) csv format when setting fields did not allow for field values that
had commas in them. switched to a proper csv parser to handle these
scenarios which was common with url values. the '=' sign similary had
issues because we were splitting on all '=' as opposed to just the first
one

(*) file-type library only works with binary file types, but common
attachments are text types. use the path and the extension in these
cases