[cli] migrate from slugs to file_paths
bump core and move away from slug management as that is a rendering
concern. also make use of simplified validation and push it to only to
the write method

in doing so, this changes all commands as they previously relied on
files OR slugs as a way to pass in a piece. that is now simplified just
to a file path.

this change also moves schemas and other configs to a single `.luzzle`
folder in the main luzzle directory.