dynamic site (#17)
since the database is changing frequently, prerendering requires rebuliding the app. rebuilding the app requires deploying the app.

as a hammer to reduce time to deploy new content, prerendering is turned off allowing the site to be fully dynamic.

this means that a caching pipeline of static assets and rendered HTML will need to be moved upstream (ex: nginx)

Reviewed-on: https://git.eleith.com/eleith/eleith/pulls/17
Co-authored-by: eleith
Co-committed-by: eleith